Renamed App to ResultsViewer
