+++ guide/results-viewer-react/src/index.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { ResultsViewer } from "./App";
+import { ResultsViewer } from "./ResultsViewer";
 import { createRoot } from "react-dom/client";
 
 const root = createRoot(document.getElementById("root")!);
